+++ src/main/java/com/todo/core/commons/model/GenericResponse.java
@@ -1,4 +1,4 @@
-package com.todo.core.config.commons.model;
+package com.todo.core.commons.model;
 
 public class GenericResponse<T> {
     public T data;
